feat: script that makes a laucher visible
Some launchers are visible only for certain desktop enviroment.

This script copies the default launcher (Desktop file) from system
applications folder '/usr/share/applications/' to local applications
folder '~/.local/share/applications/' and makes it visible for desktop
enviroments that can't "see" it normally.

The script is in the '~/.backup/script/' directory.

The old '~/.local/share/applications/lxappearance.desktop' is no longer
needed. because of this, it was removed. The

The '~/.gitignore' file was changed to remove lxappearance desktop file
and add the script.

---
Changes to be committed:
      new file:   .backup/script/update-local-applications.sh
      modified:   .gitignore
      deleted:    .local/share/applications/lxappearance.desktop