fix: system tray shadows
When closing OBS studio, stalonetray receives compton's shadows. To fix
this, the default system tray was changed to a polybar.

When this bug occurs, just right-click on the old 'tray-indicator' module
of 'simple-topbar' (now called 'sys-tray-switcher'). The new system
tray will reload and remove the shadows.

If you left-click on this module, a new stalonetray design will open and the
default system tray will be hidden.

---
Changes to be committed:
      modified:   .config/polybar/bars/simple-topbar/config.ini
      modified:   .config/polybar/bars/simple-topbar/modules.ini
      modified:   .config/polybar/bars/simple-topbar/stalonetrayrc
      new file:   .config/polybar/scripts/sys-tray.sh
      modified:   .config/polybar/update.sh