Fix: Custom polybar icon font
Fix script that installs the icon font used by polybar. Before that,
the script can't find the icon font.

This error occurred because of commit '18c1068c5d614ac10334'. This
commit renames the '~/.backup' folder to '~/.extra-files', but the script
that installs polybar icon font (inside this folder), has not been
updated.