Fix: Neovim considers a C file
If a non-C or noc-C++ file is opened, CCLS considers it to be opening a C
file. Opening C++ files thereafter will cause to CCLS to treat they as
C files.

No longer threatens non-C++ files as C files. Now considers non-C files
as C++. This does not completely fix the error, but it reduces it.